requirements.txt
request 
pytest 
pytest-playwright
pytest-html